rag_chromadb_api/requirements.txt
langchain-community
langchain-huggingface
langchain-chroma
sentence-transformers
google-genai
python-dotenv
